refactor: extract shared core logic into src/tools/core/ modules
Extract mode-agnostic business logic from tool handlers into dedicated
core modules, establishing the foundation for UI/normal mode separation.

Changes:
- Create src/tools/core/actor-execution.ts (callActorGetDataset + type)
- Create src/tools/core/actor-tools-factory.ts (getActorsAsTools, getNormalActorsAsTools, getMCPServersAsTools, enrichActorToolOutputSchemas)
- Move src/utils/actor-response.ts to src/tools/core/actor-response.ts
- Move filterRentalActors from tools/store_collection.ts to utils/actor-search.ts (fixes circular dependency)
- Update all imports; re-export from tools/actor.ts for backward compat

Pure refactor — no behavioral changes. All 248 unit tests pass.
